src/examples/langgraph/mcp_rag/requirements.txt
mcp[cli]>=1.26.0
langchain>=1.2.10
langchain-openai>=1.1.10
langgraph>=1.0.9
openai>=2.20.0
numpy>=1.26.0
python-dotenv>=1.2.1
